+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fix form select image for RTL
 - Fix scrollspy 'cls' option
 - Fix HeightMatch not working on initial page load
+- Fix touch events
 
 ## 3.0.0 beta 22 (April 24, 2017)